# Faster getPreviousMigrationId