See https://github.com/DACH-NY/cn-test-failures/issues/4500#issuecomment-2969255996

This query is currently overloading our database disk I/O.

Before:
```
QUERY PLAN
Result  (cost=0.74..0.75 rows=1 width=4) (actual time=38651.881..38651.884 rows=1 loops=1)
  Buffers: shared hit=5552 read=11529
  I/O Timings: read=25803.673
  InitPlan 1 (returns $0)
    ->  Limit  (cost=0.57..0.74 rows=1 width=4) (actual time=38651.860..38651.873 rows=1 loops=1)
          Buffers: shared hit=5552 read=11529
          I/O Timings: read=25803.673
          ->  Index Only Scan Backward using updt_hist_tran_hi_mi_rt_di on update_history_transactions  (cost=0.57..10995011.69 rows=64397161 width=4) (actual time=38651.858..38651.859 rows=1 loops=1)
                Index Cond: ((history_id = 1) AND (migration_id IS NOT NULL) AND (migration_id < 6))
                Heap Fetches: 0
                Buffers: shared hit=5552 read=11529
                I/O Timings: read=25803.673
Planning:
  Buffers: shared hit=215 read=6
  I/O Timings: read=20.340
Planning Time: 22.524 ms
Execution Time: 38652.007 ms
```

After:
```
QUERY PLAN
Aggregate  (cost=80.99..81.00 rows=1 width=4) (actual time=0.228..0.230 rows=1 loops=1)
  Buffers: shared hit=36
  CTE t
    ->  Recursive Union  (cost=0.57..78.72 rows=101 width=4) (actual time=0.066..0.216 rows=7 loops=1)
          Buffers: shared hit=36
          ->  Limit  (cost=0.57..0.74 rows=1 width=4) (actual time=0.065..0.065 rows=1 loops=1)
                Buffers: shared hit=5
                ->  Index Only Scan using updt_hist_tran_hi_mi_rt_di on update_history_transactions update_history_transactions_1  (cost=0.57..10820178.64 rows=64222991 width=4) (actual time=0.064..0.064 rows=1 loops=1)
                      Index Cond: ((history_id = 1) AND (migration_id < 6))
                      Heap Fetches: 0
                      Buffers: shared hit=5
          ->  WorkTable Scan on t t_1  (cost=0.00..7.60 rows=10 width=4) (actual time=0.020..0.020 rows=1 loops=7)
                Filter: (migration_id IS NOT NULL)
                Rows Removed by Filter: 0
                Buffers: shared hit=31
                SubPlan 1
                  ->  Limit  (cost=0.57..0.74 rows=1 width=4) (actual time=0.021..0.021 rows=1 loops=6)
                        Buffers: shared hit=31
                        ->  Index Only Scan using updt_hist_tran_hi_mi_rt_di on update_history_transactions  (cost=0.57..62412.73 rows=362596 width=4) (actual time=0.020..0.020 rows=1 loops=6)
                              Index Cond: ((history_id = 1) AND (migration_id < 6) AND (migration_id > t_1.migration_id))
                              Heap Fetches: 0
                              Buffers: shared hit=31
  ->  CTE Scan on t  (cost=0.00..2.02 rows=100 width=4) (actual time=0.069..0.222 rows=6 loops=1)
        Filter: (migration_id IS NOT NULL)
        Rows Removed by Filter: 1
        Buffers: shared hit=36
Planning:
  Buffers: shared hit=213
Planning Time: 2.026 ms
Execution Time: 0.609 ms
```